Rewrite the Rustone core for modularity
Preparing for Forge/vanilla/Nukkit/etc. server support. Server types are
modules in the code, called kinds. One rustone::server_kinds::ServerKind
enum incorporates these modules. Functions in these modules initialize,
launch the servers, download jars, and do other type-specific stuff.

While these modules provide helper functions, actually executing the
download is the responsibility of the frontend; this allows rscmd to
display a fancy progress bar, and rshttp to signal whether it needs to
upgrade or download a jar file to the request invoker.

At the moment, only Paper servers are supported - the name is 'paper'.
Forge support is planned to be added soon.

`rscmd create` has a new parameter: --type or -T, this indicates the type
of the server to be created with its name.

rshttp returns HTTP 202 ACCEPTED when it needs to download or upgrade a
jar before starting the server, and uses HTTP 404 NOT FOUND to signal
the absence of a server.

rscmd shows progress bars when downloading or upgrading a server jar.